% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00augmentCov.R
\name{augmentCov}
\alias{augmentCov}
\title{Augment a covariance matrix}
\usage{
augmentCov(
  covmat,
  center,
  dnames = list(c("int", colnames(covmat)), c("int", colnames(covmat)))
)
}
\arguments{
\item{covmat}{covariance matrix of a dataset}

\item{center}{numeric vector of means of a dataset}

\item{dnames}{A list of dimension names (if they should be changed from the names of the columns of covmat)}
}
\value{
The original covariance matrix augmented with the vector of means.
}
\description{
Adds the vector of means as a column and row to any covariance matrix
}
\details{
This function only combines a vector of means and a covariance matrix into an augmented covariance matrix, it does not produce any new estimate.
}
\examples{
# Run the function
augmentCov(
  covmat = cov(mtcars), # covariance matrix of a dataset
  center = colMeans(mtcars) # vector of means of a dataset
)

}
\author{
Edoardo Costantini, 2024
}
