% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputPlot_modelselection_CHull.r
\name{chull_lmfa}
\alias{chull_lmfa}
\title{CHull}
\usage{
chull_lmfa(modelselection, PercentageFit = 0.01, ...)
}
\arguments{
\item{modelselection}{min.}

\item{PercentageFit}{Required proportion of increase in fit of a more complex model}

\item{...}{rows.}
}
\value{
Returns
}
\description{
This fun
}
\examples{
\dontrun{

}
}
