% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00createSOMI.R
\name{createSOMI}
\alias{createSOMI}
\title{Create a SOMI object}
\usage{
createSOMI(X)
}
\arguments{
\item{X}{A rectangular data set (matrix or data.frame) with named columns.}
}
\value{
Object of class \code{somi}, which is a list containing:
 \itemize{
   \item \code{S}: unit integer vector storing the number of missing data patterns
   \item \code{O}: list of names of fully observed variables in each missing data pattern
   \item \code{M}: list of names of variables with missing values in each missing data pattern
   \item \code{I}: list of membership of data rows in each missing data pattern
}
}
\description{
\code{createSOMI} stores the names of the fully observed variables, the names of the variables with missing values, the identifiers of the data rows belonging to each missing data pattern, for each missing data pattern in the input dataset.
}
\examples{
# Set a seed
set.seed(100)

# Define an input matrix with simple known missing data
data_matrix <- matrix(
  c(
    0, 1, NA,
    1, 2, NA,
    NA, 2, 3,
    5, NA, NA
  ),
  ncol = 3,
  dimnames = list(NULL, paste0("x", 1:3)),
  byrow = TRUE
)

# Use the function
SOMI_object <- createSOMI(X = data_matrix)
}
\author{
Edoardo Costantini, 2023
}
