% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8ESMdata.R
\docType{data}
\name{ESM}
\alias{ESM}
\title{Example ESM dataset}
\format{
A data frame with 4776 rows and 23 variables:
\itemize{
  \item id:
  \item deltaT:
  \item negativeEvent:
  \item intervention:
  \item interactions:
  \item Interested:
  \item Joyful:
  \item Determined:
  \item Calm:
  \item Lively:
  \item Enthusiastic:
  \item Relaxed:
  \item Cheerful:
  \item Content:
  \item Energetic:
  \item Upset:
  \item Gloomy:
  \item Sluggish:
  \item Anxious:
  \item Bored:
  \item Irritated:
  \item Nervous:
  \item Listless:
}
}
\usage{
ESM
}
\description{
The variables are as follows:
}
\keyword{datasets}
