% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step1Factorscores.r
\name{factorscores_lmfa}
\alias{factorscores_lmfa}
\title{Caclulating initial state and transition probabilities}
\usage{
factorscores_lmfa(data, model, oblique = TRUE, rounding = 4)
}
\arguments{
\item{data}{The dataset (must be a dataframe and contain complete cases only).}

\item{model}{The model estimated with \code{step1} (must be of class lmfa_step1).}

\item{oblique}{Indicates whether the factor scores are obtained for the obliquely rotated loadings (TRUE) or unrotated loadings (FALSE) (must be a logical statement).}

\item{rounding}{The number of decimals to which the results should be rounded (must be a single scalar).}
}
\description{
\code{probabilities} calculates initial state and transition probabilities for given covariate scores and time interval for the \code{step3} estimates.
}
\examples{
\dontrun{
ESM_factorscores <- factorscores_lmfa(data, 
                     model, 
                     oblique = TRUE,
                     rounding = 4
                     )
}
}
