% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_NA.R
\name{initialize_NAs}
\alias{initialize_NAs}
\title{Initialize missing values}
\usage{
initialize_NAs(x, replace = TRUE)
}
\arguments{
\item{x}{A vector with missing values}

\item{replace}{A boolean unit vector describing whether sampling should be done with replacement: yes (TRUE), or no (FALSE)}
}
\value{
A vector with missing values replaced by sampled observed values
}
\description{
Initialize missing values on a data column. If there are no missing values, the original vector is returned.
}
\examples{
# Create a data matrix
x <- c(1, 5, 2, 4, NA, 2, 5, NA)

# Use function
initialize_NAs(x)

}
\author{
Edoardo Costantini, 2023
}
