% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1InitializeEM.R
\name{initializeStep1}
\alias{initializeStep1}
\title{Initialize all Parameters}
\usage{
initializeStep1(x, n_sub, n_state, n_fact, J, startval = "random")
}
\arguments{
\item{x}{The dataset.}

\item{n_sub}{Number of observations.}

\item{n_state}{Number of states. Has to be a scalar.}

\item{n_fact}{Number of factors for each state. Has to be a vector of length n_state.}

\item{J}{Number of items.}

\item{startval}{Indicating whether startvalues are based on random assignment or mclust.}
}
\value{
Returns list with all the state-specific initialized parameters.
}
\description{
Initial state-membership probabilities are based on mclust and random assignment.
}
\examples{
\dontrun{
InitialValues <-initializeStep1(
x = x, 
n_sub = 500, 
n_state = 2, 
n_fact = c(2,2), 
J = 20,
startval = "random")
z_ik<- InitialValues$z_ik
N_k<- InitialValues$N_k
pi_k<- InitialValues$pi_k
mu_k<- InitialValues$mu_k
C_k<- InitialValues$C_k
Lambda_k<- InitialValues$Lambda_k
Psi_k<- InitialValues$Psi_k
}
}
