% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step3Invariance.r
\name{invariance}
\alias{invariance}
\title{Evaluating for which individuals invariance holds across all measurement occasions}
\usage{
invariance(model, identifier)
}
\arguments{
\item{model}{Output from the main function step3() (must be of class lmfa_step3).}

\item{identifier}{The name of the column containing the subject identifiers (must be a single character).}
}
\description{
\code{invariance} evaluates for which individuals invariance holds across all measurement occasions based on the \code{step3} output.
}
\examples{
\dontrun{
invariance(model, identifier)
}
}
