% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lmfa}
\alias{lmfa}
\alias{lmfa-package}
\title{`lmfa`: The R-package for exploring within-person changes and between person differences in measurement models in (intensive) longitudinal data.}
\description{
What the package does (one paragraph).
}
\details{
The package employs a three-step estimation of continuous-time latent Markov factor analysis.
}
\section{Section name}{

You can add multiple sections.
}

\author{
\strong{Maintainer}: Leonie Vogelsmeier \email{l.v.d.e.vogelsmeier@uvt.nl} (YOUR-ORCID-ID)

}
