% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8modelselection.R
\docType{data}
\name{modelselection}
\alias{modelselection}
\title{Model selection results for the example ESM data for 1-4 states and 2-3 factors per state}
\format{
list
}
\usage{
modelselection
}
\description{
Model selection results for the example ESM data for 1-4 states and 2-3 factors per state
}
\keyword{datasets}
