% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_LL.R
\name{observed_LL}
\alias{observed_LL}
\title{Compute observed data log-likelihood}
\usage{
observed_LL(dmv, n, pi_k)
}
\arguments{
\item{dmv}{The DMV coefficients produced by function DMV}

\item{n}{Number of observations.}

\item{pi_k}{The state-specific loading matrices.}
}
\value{
Observed data log-likelihood
}
\description{
Compute observed data log-likelihood
}
