% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PlotChull.r
\name{plot.CHull}
\alias{plot.CHull}
\title{Plot function for: A Generic Convex-Hull-Based Model Selection Method}
\usage{
\method{plot}{CHull}(x, col = NULL, pch = NULL, ...)
}
\arguments{
\item{x}{Output from the main function step1() (must be of class lmfa_modelselection).}

\item{col}{Vector of \code{\link{colors}} used for plots.}

\item{pch}{Symbol used to indicate selected model(s).}

\item{...}{Further arguments for the CHull plot function.}
}
\description{
This function is based on the CHull function from the R package multichull
}
