% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputPlot_modelselection.r
\name{plot.lmfa_modelselection}
\alias{plot.lmfa_modelselection}
\title{Plot for model selection}
\usage{
\method{plot}{lmfa_modelselection}(x, ...)
}
\arguments{
\item{x}{An x storing output from the main function Step3()}

\item{...}{Further arguments for the default S3 summary method}
}
\description{
Plot for model selection
}
\examples{
\dontrun{
plot(results_modelselection)
}
}
