% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step1.R
\name{step1}
\alias{step1}
\title{Estimating state-specific measurement models}
\usage{
step1(
  data,
  indicators,
  n_state = NULL,
  n_fact = NULL,
  modelselection = FALSE,
  n_state_range = NULL,
  n_fact_range = NULL,
  n_starts = 25,
  n_initial_ite = 15,
  n_m_step = 10,
  em_tolerance = 1e-08,
  m_step_tolerance = 0.001,
  max_iterations = 1000,
  n_mclust = 5
)
}
\arguments{
\item{data}{The dataset (must be a dataframe and contain complete cases only).}

\item{indicators}{The variable names of the indicators (must be a vector of characters).}

\item{n_state}{The number of states that should be estimated (must be a single scalar).}

\item{n_fact}{The number of factors per state that should be estimated (must be a numeric vector of length n_state).}

\item{modelselection}{Indicates whether model selection should be performed or not. If TRUE, the arguments n_state_range and n_fact_range are required (must be a logical statement).}

\item{n_state_range}{A vector indicating the number of states that should be considered in the model selection.}

\item{n_fact_range}{A vector indicating the number of factors that should be considered in the model selection.}

\item{n_starts}{The number of random starts that should be used (must be a single scalar).}

\item{n_initial_ite}{The number of initial iterations for the best starts (must be a single scalar).}

\item{n_m_step}{The number of M-step iterations that should be used when parameters still change more than defined by the m_step_tolerance (must be a single scalar).}

\item{em_tolerance}{The convergence criterion for parameters and loglikelihood (must be a single scalar and smaller than m_step_tolerance).}

\item{m_step_tolerance}{The criterion for stopping the n_m_step M-step iterations (must be a single scalar).}

\item{max_iterations}{The maximum number of iterations (must be a single scalar and larger than n_initial_ite).}

\item{n_mclust}{The number of mclust starts (must be a single scalar and at least equal to 2).}
}
\value{
Returns the state-specific measurement model parameters and model fit information (for one or multiple estimated model).
}
\description{
\code{step1} conducts step 1 of the three-step estimation of LMFA and thus the estimation of the measurement models. It is possible to estimate the parameters for a single model or for a range of models and to conduct model selection.
}
\examples{
\dontrun{
step1_results <- step1(data,
  indicators,
  n_state = NULL,
  n_fact = NULL,
  modelselection = FALSE,
  n_state_range = NULL,
  n_fact_range = NULL,
  n_starts = 25,
  n_initial_ite = 15,
  n_m_step = 10,
  em_tolerance = 1e-8,
  m_step_tolerance = 1e-3,
  max_iterations = 1000,
  n_mclust = 5
)
}
}
