% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step2.r
\name{step2}
\alias{step2}
\title{Obtaining state assignment and classification errors}
\usage{
step2(data, model)
}
\arguments{
\item{data}{The dataset (must be a dataframe and contain complete cases only).}

\item{model}{The model estimated with \code{step1} (must be of class lmfa_step1).}
}
\value{
Returns the posterior state probabilities, the modal state assignments, the classification errors, and the state proportions.
}
\description{
\code{step2} conducts step 2 of the three-step estimation of LMFA. More specifically, the function extracts the classification information from the \code{step1} output.
}
\examples{
\dontrun{
classification <- step2(data, model)
}
}
