% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step3.R
\name{Step3}
\alias{Step3}
\title{Conducts step 3 from the three-step estimation of CT-LMFA.}
\usage{
Step3(input_file, interval_column, id_column, n_state, fitStep1Step2,
  transitionCovariates = NULL, initialCovariates = NULL,
  i.method = "BFGS", i.maxit = 10000, i.reltol = 1e-08, n_q = 10,
  n_initial_ite = 10, previousCov = FALSE)
}
\arguments{
\item{input_file}{The dataset. Should contain complete cases only 
(unequal intervals are automatically dealt with in the CT-LMM).}

\item{interval_column}{The column number or name with intervals.}

\item{id_column}{The column number or name with subject-id.}

\item{n_state}{The number of states that should be estimated 
(has to be a scalar).}

\item{fitStep1Step2}{The output file created with Step1Step2().}

\item{transitionCovariates}{Vector with covariates for the transition intensities.}

\item{initialCovariates}{Vector with covariates for the initial state probabilities.}

\item{i.method}{N}

\item{i.maxit}{N}

\item{i.reltol}{N}

\item{n_q}{N}

\item{n_initial_ite}{N}

\item{previousCov}{N}
}
\value{
Returns .
}
\description{
Conducts step 3 from the three-step estimation of CT-LMFA.
}
\examples{
\dontrun{
step3Results <- Step3(input_file,
interval_column,
id_column,
n_state,
fitStep1Step2,
transitionCovariates = NULL,
initialCovariates = NULL,
i.method = "BFGS",
i.maxit = 10000,
i.reltol = 1e-8,
n_q = 10,
n_initial_ite = 10,
previousCov = FALSE)
}
}
