% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputSummary.r
\name{summary.lmfa}
\alias{summary.lmfa}
\title{Summary}
\usage{
\method{summary}{lmfa}(object, ...)
}
\arguments{
\item{object}{The dataset (must be a dataframe and contain complete cases only).}
}
\value{
Returns the measurement model parameters, the proportional and
modal state assignments, and the classification errors.
}
\description{
Summary
}
\examples{
\dontrun{
fitStep1Step2 <- Step1Step2(input_file,variable_columns,n_state,
                      n_fact,n_starts=25,n_initial_ite=15,n_m_step=10,
                      em_tolerance=1e-6,m_step_tolerance=1e-3,max_iterations=500)
}
}
