% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0SummaryStep1CHull.r
\name{summary.lmfa_chull}
\alias{summary.lmfa_chull}
\title{Summary for \code{lmfa_chull}}
\usage{
\method{summary}{lmfa_chull}(object, ...)
}
\arguments{
\item{object}{An object storing output from \code{lmfa_chull}.}

\item{...}{Further arguments for the default S3 summary method.}
}
\description{
Summary for \code{lmfa_chull}
}
\examples{
\dontrun{
summary(chullresults)
}
}
