% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputSummary_modelselection.r
\name{summary.lmfa_modelselection}
\alias{summary.lmfa_modelselection}
\title{Summary statistics for modelselection}
\usage{
\method{summary}{lmfa_modelselection}(object, ...)
}
\arguments{
\item{object}{An object storing output from the main function Step1Step2()}

\item{...}{Further arguments for the default S3 summary method}
}
\description{
Summary statistics for modelselection
}
\examples{
\dontrun{
summary(results1)
}
}
