% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputSummaryProbabilities.r
\name{summary.lmfa_prob}
\alias{summary.lmfa_prob}
\title{Summary for probabilities()}
\usage{
\method{summary}{lmfa_prob}(object, ...)
}
\arguments{
\item{object}{An object storing output from the probabilities() function.}

\item{...}{Further arguments for the default S3 summary method.}
}
\description{
Summary for probabilities()
}
\examples{
\dontrun{
summary(results3)
}
}
