% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputSummary.r
\name{summary.lmfa_step1step2}
\alias{summary.lmfa_step1step2}
\title{Summary statistics for Step1Step2 analysis}
\usage{
\method{summary}{lmfa_step1step2}(object, ...)
}
\arguments{
\item{object}{An object storing output from the main function Step1Step2()}

\item{...}{Further arguments for the default S3 summary method}
}
\description{
Summary statistics for Step1Step2 analysis
}
\examples{
\dontrun{
summary(results1)
}
}
