% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputSummaryStep2.r
\name{summary.lmfa_step2}
\alias{summary.lmfa_step2}
\title{Summary statistics for step2()}
\usage{
\method{summary}{lmfa_step2}(object, ...)
}
\arguments{
\item{object}{An object storing output from the main function step2()}

\item{...}{Further arguments for the default S3 summary method}
}
\description{
Summary statistics for step2()
}
\examples{
\dontrun{
summary(results2)
}
}
