% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputSummaryStep3.r
\name{summary.lmfa_step3}
\alias{summary.lmfa_step3}
\title{Summary statistics for step3()}
\usage{
\method{summary}{lmfa_step3}(object, ...)
}
\arguments{
\item{object}{An object storing output from the main function Step3()}

\item{...}{Further arguments for the default S3 summary method}
}
\description{
Summary statistics for step3()
}
\examples{
\dontrun{
summary(results3)
}
}
