% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{sweep}
\alias{sweep}
\title{Sweep operator}
\usage{
sweep(A, target)
}
\arguments{
\item{A}{Symmetric matrix}

\item{target}{Integer vector indicating which columns of A should be swept in.}
}
\value{
The H matrix resulting from the sweeping: H = SWP[target]A
}
\description{
This function implements the sweep operator described by Goodnight 1979, p. 154.
}
\details{
Some details on what this function does.
}
\author{
Edoardo Costantini, 2021
}
