% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8transitionmodel.R
\docType{data}
\name{transitionmodel}
\alias{transitionmodel}
\title{Transition model results for the example ESM data with two covariates}
\format{
list
}
\usage{
transitionmodel
}
\description{
Transition model results for the example ESM data with two covariates
}
\keyword{datasets}
