% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00updateTmat.R
\name{updateTmat}
\alias{updateTmat}
\title{Create a SOMI object}
\usage{
updateTmat(X, wt = rep(1, ncol(X)), Tmat, theta, obs, v_mis, v_obs)
}
\arguments{
\item{X}{A rectangular data set (matrix or data.frame) with named columns.}

\item{wt}{A vector of weights for each column}

\item{Tmat}{Original matrix of sufficient statistics to update}

\item{theta}{Augmented covariance matrix as produced by \code{augmentCov}}

\item{obs}{numeric vector indicating of membership of data rows to the missing data pattern}

\item{v_mis}{character vector of names of variables with missing values the missing data pattern}

\item{v_obs}{character vector of names of fully observed variables in the missing data pattern}
}
\value{
an updated Tmat (matrix)
}
\description{
Updates the matrix of sufficient statistics for an E step for a given missing data pattern
}
\examples{
# Create a data matrix
X <- matrix(data = c(16,20,16,20,-6,-4, 12,24,12,-6,4,-8,
                       8,8,26,-4,4,8, 20,8,NA,NA,20,-4,
                       8,4,-8,NA,22,-8, 10,20,28,-20,-4,-4,
                       4,28,24,12,8,18, -8,20,24,-3,8,-24,
                       NA,20,24,8,12,NA),
              ncol = 6,
              byrow = TRUE,
              dimnames = list(NULL, c("15P","15L","15H", "90P", "90L", "90H")))

# Create a SOMI object for this matrix
SOMI <- createSOMI(X)

# Compute starting matrices
Tmat = computeTobs(X, wt = wt, S = SOMI$S, I = SOMI$I)
theta = augmentCov(
  covmat = cov(X, use = "complete.obs"),
  center = colMeans(X, na.rm = TRUE)
) # current augmented covariance matrix

# Update the Tmat with info in the forth missing data pattern
updateTmat(
    X = X,
    wt = wt,
    Tmat = Tmat,
    theta = theta,
    obs = SOMI$I[[4]],
    v_mis = SOMI$M[[4]],
    v_obs = SOMI$O[[4]]
)

}
\author{
Edoardo Costantini, 2023
}
