% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0OutputPlot_modelselection_CHull.r
\name{chull_lmfa}
\alias{chull_lmfa}
\title{A Generic Convex-Hull-Based Model Selection Method}
\usage{
chull_lmfa(x, PercentageFit = 0.01, ...)
}
\arguments{
\item{x}{Output from the main function step1() (must be of class lmfa_step1).}

\item{PercentageFit}{Required proportion of increase in fit of a more complex model.}

\item{...}{Further arguments for the CHull plot function.}
}
\value{
Returns CHull output.
}
\description{
This function is based on the CHull function from the R package multichull
}
\examples{
\dontrun{
chull_lmfa(x)
}
}
