% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step1Chull.r
\name{chull_lmfa}
\alias{chull_lmfa}
\title{A Generic Convex-Hull-Based Model Selection Method}
\usage{
chull_lmfa(x, PercentageFit = 0.001, ...)
}
\arguments{
\item{x}{Model-selection output from \code{step1} (must be of class lmfa_modelselection).}

\item{PercentageFit}{Required proportion of increase in fit of a more complex model.}

\item{...}{Further arguments for the CHull plot function.}
}
\value{
Returns CHull output.
}
\description{
\code{chull_lmfa} is based on \code{CHull} from the R package multichull.
}
\examples{
\dontrun{
chull_lmfa(x)
}
}
