% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lmfa}
\alias{lmfa}
\alias{lmfa-package}
\title{`lmfa`: The R-package for exploring within-person changes and between person differences in measurement models in (intensive) longitudinal data.}
\description{
`lmfa` allows to explore within-person changes 
    and between person differences in measurement models in 
    (intensive) longitudinal data by means of three-step 
    continuous-time latent Markov factor analysis.
}
\details{
The package employs a three-step estimation of continuous-time latent Markov factor analysis.
}
\section{Section name}{

You can add multiple sections.
}

\author{
\strong{Maintainer}: Leonie V. D. E. Vogelsmeier \email{l.v.d.e.vogelsmeier@uvt.nl} (\href{https://orcid.org/0000-0002-1666-7112}{ORCID})

Authors:
\itemize{
  \item Kim De Roover \email{K.DeRoover@tilburguniversity.edu}
}

}
