% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PlotBIC.r
\name{plot.lmfa_modelselection}
\alias{plot.lmfa_modelselection}
\title{Plot for model selection}
\usage{
\method{plot}{lmfa_modelselection}(x, ...)
}
\arguments{
\item{x}{Output from the main function step1() (must be of class lmfa_modelselection).}

\item{...}{Further arguments for the default S3 plot method.}
}
\description{
Plot for model selection
}
\examples{
\dontrun{
plot(modelselection)
}
}
