% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step2.r
\name{step2}
\alias{step2}
\title{Conducts steps 1 and 2 from the three-step estimation of CT-LMFA.}
\usage{
step2(data, model)
}
\arguments{
\item{data}{The dataset (must be a dataframe and contain complete cases only).}

\item{model}{The model estimated with step1() (must be of class lmfa_step1).}
}
\value{
Returns the measurement model parameters, the proportional and
modal state assignments, and the classification errors.
}
\description{
Conducts steps 1 and 2 from the three-step estimation of CT-LMFA.
}
\examples{
\dontrun{
step2_results <- step2(data, model)
}
}
