% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step3.R
\name{Step3}
\alias{Step3}
\title{Conducts step 3 from the three-step estimation of CT-LMFA.}
\usage{
Step3(
  input_file,
  id_column,
  n_state,
  post.probabilities,
  interval_column = NULL,
  transitionCovariates = NULL,
  initialCovariates = NULL,
  i.method = "BFGS",
  i.maxit = 10000,
  i.reltol = 1e-08,
  n_q = 10,
  n_initial_ite = 10,
  previousCov = FALSE
)
}
\arguments{
\item{input_file}{The dataset (must be a dataframe).}

\item{id_column}{The name with subject identifiers (must be a single character).}

\item{n_state}{The number of states that should be estimated (must be a single scalar).}

\item{post.probabilities}{The posterior state-membership probabilities (must be a dataframe with n_state columns and of same length as the input_file).}

\item{interval_column}{The column name with intervals (must be a single character).}

\item{transitionCovariates}{The covariate(s) for the transition intensities (must be a (vector of) character(s)).}

\item{initialCovariates}{The covariate(s) for the initial state probabilities (must be a (vector of) character(s)).}

\item{i.method}{The type of optimization method that should be used (must be "BFGS" or "CG")}

\item{i.maxit}{The maximum number of iterations that should be used (must be a single scalar and larger than n_initial_ite).}

\item{i.reltol}{The tolerance to evaluate convergend that should be used (must be a single scalar).}

\item{n_q}{The number of start values for the transition intensity parameters that should be used (must be a single scalar).}

\item{n_initial_ite}{The number of initial iterations for the different start sets that should be used (must be a single scalar).}

\item{previousCov}{Indicates whether the covariate at t (FALSE) or t-1 (TRUE) should be used (previousCov must be a single logical statement).}
}
\value{
Returns .
}
\description{
Conducts step 3 from the three-step estimation of CT-LMFA.
}
\examples{
\dontrun{
step3Results <- Step3(input_file,
interval_column,
id_column,
fitStep1Step2,
transitionCovariates = NULL,
initialCovariates = NULL,
i.method = "BFGS",
i.maxit = 10000,
i.reltol = 1e-8,
n_q = 10,
n_initial_ite = 10,
previousCov = FALSE)
}
}
