% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step3.R
\name{step3}
\alias{step3}
\title{Estimating the transition model (with or without covariates) by means of a CT-LMM}
\usage{
step3(
  data,
  identifier,
  n_state,
  postprobs,
  timeintervals = NULL,
  initialCovariates = NULL,
  transitionCovariates = NULL,
  n_starts = 25,
  n_initial_ite = 10,
  method = "BFGS",
  max_iterations = 10000,
  tolerance = 1e-10,
  scaling = "proxi"
)
}
\arguments{
\item{data}{The dataset (must be a dataframe).}

\item{identifier}{The name of the column containing the subject identifiers (must be a single character).}

\item{n_state}{The number of states that should be estimated (must be a single scalar).}

\item{postprobs}{The posterior state-membership probabilities (must be a dataframe with n_state columns and of same length as the data).}

\item{timeintervals}{The name of the column containing the intervals (must be a single character).}

\item{initialCovariates}{The covariate(s) for the initial state probabilities (must be a (vector of) character(s)).}

\item{transitionCovariates}{The covariate(s) for the transition intensities (must be a (vector of) character(s)).}

\item{n_starts}{The number of start values for the transition intensity parameters that should be used (must be a single scalar).}

\item{n_initial_ite}{The number of initial iterations for the different start sets that should be used (must be a single scalar).}

\item{method}{The type of optimization method that should be used (must be "BFGS" or "CG")}

\item{max_iterations}{The maximum number of iterations that should be used (must be a single scalar and larger than n_initial_ite).}

\item{tolerance}{The tolerance to evaluate convergend that should be used (must be a single scalar).}

\item{scaling}{An overall scaling to be applied to the value of fn (a function to be minimized) and gr (a function to return the gradient for the "BFGS" and "CG" methods) during optimization (see optim() docomentation for details). In this package it has to be a positive integer.}
}
\value{
Returns the transition model parameters.
}
\description{
\code{step3} conducts step 3 from the three-step estimation of CT-LMFA and thus the estimation of the transition model. To this end, the function uses the classification information from the \code{step2} output. Makes use of \code{msm} from the msm package.
}
\examples{
\dontrun{
step3_results <- step3(data,
                 identifier,
                 n_state,
                 postprobs,
                 timeintervals = NULL,
                 transitionCovariates = NULL,
                 initialCovariates = NULL,
                 n_starts = 25,
                 n_initial_ite = 10,
                 method = "BFGS",
                 max_iterations = 10000,
                 tolerance = 1e-10,
                 scaling = 1
                 )
}
}
