% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7Step3.R
\name{Step3}
\alias{Step3}
\title{Conducts step 3 from the three-step estimation of CT-LMFA.}
\usage{
Step3(
  data,
  identifier,
  n_state,
  post.probabilities,
  timeintervals = NULL,
  transitionCovariates = NULL,
  initialCovariates = NULL,
  i.method = "BFGS",
  i.maxit = 10000,
  i.reltol = 1e-08,
  i.fnscale = "proxi",
  n_q = 5,
  n_initial_ite = 10
)
}
\arguments{
\item{data}{The dataset (must be a dataframe).}

\item{identifier}{The name with subject identifiers (must be a single character).}

\item{n_state}{The number of states that should be estimated (must be a single scalar).}

\item{post.probabilities}{The posterior state-membership probabilities (must be a dataframe with n_state columns and of same length as the data).}

\item{timeintervals}{The column name with intervals (must be a single character).}

\item{transitionCovariates}{The covariate(s) for the transition intensities (must be a (vector of) character(s)).}

\item{initialCovariates}{The covariate(s) for the initial state probabilities (must be a (vector of) character(s)).}

\item{i.method}{The type of optimization method that should be used (must be "BFGS" or "CG")}

\item{i.maxit}{The maximum number of iterations that should be used (must be a single scalar and larger than n_initial_ite).}

\item{i.reltol}{The tolerance to evaluate convergend that should be used (must be a single scalar).}

\item{i.fnscale}{An overall scaling to be applied to the value of fn (a function to be minimized) and gr (a function to return the gradient for the "BFGS" and "CG" methods) during optimization (see optim() docomentation for details). In this package it has to be a positive integer.}

\item{n_q}{The number of start values for the transition intensity parameters that should be used (must be a single scalar).}

\item{n_initial_ite}{The number of initial iterations for the different start sets that should be used (must be a single scalar).}
}
\value{
Returns .
}
\description{
Conducts step 3 from the three-step estimation of CT-LMFA.
}
\examples{
\dontrun{
step3Results <- Step3(data,
                 identifier,
                 n_state,
                 post.probabilities,
                 timeintervals = NULL,
                 transitionCovariates = NULL,
                 initialCovariates = NULL,
                 i.method = "BFGS",
                 i.maxit = 10000,
                 i.reltol = 1e-8,
                 i.fnscale = 1,
                 n_q = 5,
                 n_initial_ite = 10
                 )
}
}
