% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0SummaryStep1Modelselection.r
\name{summary.lmfa_modelselection}
\alias{summary.lmfa_modelselection}
\title{Summary statistics for \code{step1} with model selection}
\usage{
\method{summary}{lmfa_modelselection}(object, ...)
}
\arguments{
\item{object}{An object storing output from \code{step1} with model selection.}

\item{...}{Further arguments for the default S3 summary method.}
}
\description{
Summary statistics for \code{step1} with model selection
}
\examples{
\dontrun{
summary(results1)
}
}
