% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0SummaryProbabilities.r
\name{summary.lmfa_prob}
\alias{summary.lmfa_prob}
\title{Summary for \code{probabilities}}
\usage{
\method{summary}{lmfa_prob}(object, ...)
}
\arguments{
\item{object}{An object storing output from \code{probabilities}.}

\item{...}{Further arguments for the default S3 summary method.}
}
\description{
Summary for \code{probabilities}
}
\examples{
\dontrun{
summary(results3)
}
}
