% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0SummaryStep1.r
\name{summary.lmfa_step1}
\alias{summary.lmfa_step1}
\title{Summary statistics for \code{step1}}
\usage{
\method{summary}{lmfa_step1}(object, rounding = 2, ...)
}
\arguments{
\item{object}{An object storing output from \code{step1}.}

\item{rounding}{The number of decimals to which the results should be rounded (must be a single scalar).}

\item{...}{Further arguments for the default S3 summary method}
}
\description{
Summary statistics for \code{step1}
}
\examples{
\dontrun{
summary(measurementmodel)
}
}
