% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0SummaryStep3.r
\name{summary.lmfa_step3}
\alias{summary.lmfa_step3}
\title{Summary statistics for \code{step3}}
\usage{
\method{summary}{lmfa_step3}(object, rounding = 4, ...)
}
\arguments{
\item{object}{An object storing output from \code{step3} (must be of class lmfa_modelselection).}

\item{rounding}{The number of decimals to which the results should be rounded (must be a single scalar).}

\item{...}{Further arguments for the default S3 summary method.}
}
\description{
Summary statistics for \code{step3}
}
\examples{
\dontrun{
summary(transitionmodel)
}
}
